perf: int16 quantization (det 1874→2688+) + drop the per-point id array
Offline diagnosis vs the float oracle over the full 3M set showed the int8 store
was the last detection gap: int8 flips ~68 borderline k-NN decisions over the
54100 official queries; float64 (the oracle's own metric) flips zero. Switching
the store to int16 (QUANT_SCALE 32766, signed short) makes quantization error
negligible — exact-search det 1874→2819, budget 4096 det 1859→2688, budget 2048
1584→2155.

int16 doubles the store to 84MB, so to stay inside the 160MiB/instance budget
the per-point id array is gone: points are packed contiguously per bucket in
tree-traversal order (VpTree reorders its store slice in place at build), so a
point's id is bucketBase+position. Index is 92MB on disk / ~97MB heap (was
OOM-tight at ~110MB with ids); loaded in chunks to avoid a 2nd full-size buffer.
The contiguous layout keeps search reads cache-local: ns/comp only rose 42→52
despite the 2x larger store. Magic bumped FSI1→FSI2.

Tests green (round-trip + exactness vs the quantized and float oracles).

+++ src/test/kotlin/QuantizedBruteForceIndex.kt
@@ -19,7 +19,7 @@ class QuantizedBruteForceIndex(
     references: List<LabeledVector>,
     private val dim: Int = VECTOR_DIMENSIONS,
 ) : VectorIndex {
-    private val store = ByteArray(references.size * dim)
+    private val store = ShortArray(references.size * dim)
     private val labels = BooleanArray(references.size)
 
     init {
